+++ resources/views/layouts/app.blade.php
@@ -10,6 +10,8 @@
     <link rel="stylesheet" href="{{ URL::asset('css/sidebar.css') }}">
 
     @vite(['resources/css/app.css', 'resources/js/app.js'])
+
+    @livewireStyles
 </head>
 <body class="font-sans antialiased">
 <div class="min-h-screen bg-gray-100">
@@ -27,5 +29,7 @@
         {{ $slot }}
     </main>
 </div>
+
+@livewireScripts
 </body>
 </html>